factor out a lot of generic 'engine-y' stuff from the coup game and put all that game specific stuff in new example directory
